requirements.txt
aiosqlite==0.12.0
asyncpg==0.20.1
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
ciso8601==2.1.3
click==7.1.1
databases==0.2.6
dnspython==1.16.0
email-validator==1.0.5
envparse==0.2.0
fastapi==0.54.1
fastapi-users==0.7.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.4.21
httptools==0.1.1
httpx==0.12.1
httpx-oauth==0.2.2
hyperframe==5.2.0
idna==2.9
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy==0.770
mypy-extensions==0.4.3
passlib==1.7.2
psycopg2-binary==2.8.5
pycparser==2.20
pydantic==1.5
PyJWT==1.7.1
PyPika==0.37.3
python-multipart==0.0.5
rfc3986==1.4.0
six==1.14.0
sniffio==1.1.0
SQLAlchemy==1.3.16
starlette==0.13.2
tortoise-orm==0.16.10
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
